Update Javadoc for FileReferenceStore.splitFileReferences
